Implement the first global signal
Global signals are sent to all main windows.
The signal is fiat_ccy_changed and the method called is
on_fiat_ccy_changed.  Unfortunately implementing this
caused a cascade of changes as the main window no
longer has an fx object, so the change is not as large
as it appears.

Previously changing the fiat currency in preferences caused
only one window to be updated (the parent of the preferences
dialog).  Now they all update.

There are *many* more similar cases.  Also there are many
redundant refreshes of the main window, which should be
revisited.