Migration to full sqlite storage. All unit tests pass. Close #213.
Fix the broken reorg handling. The lock needed to be reentrant. Switched from filtering based on the relevant metadata fields to the more definitive 'cleared' state. Updated the cache entries to match the existing update to the store. Added a unit test to test the reorg changes were applied, including edge conditions. Close #229.