This commit represents the presence of a bug in our networking code, or the dependencies it uses.  The upgrade from asyncio 0.19 to 0.22.1 changed the behaviour of task groups and any cancelled exception within now causes the whole group to be cancelled. Unfortunately, what is getting cancelled where is opaque and hard to track down.  We originally observed that when an ElectrumSV instance had multiple wallets windows open and one was closed, the network would shut down because the "remove wallet" code in the network cancelled the task managing that wallet (that cancel in turn cancelled all remaining tasks in the taskgroup for the network main task).
During the recent reorg event, I observed that my network main task exited.  Given that taskgroups swallow all exceptions, it is impossible to know why. This change does two things. It converts the main task into a main loop that restarts the networking if it exits, unless the network is in the process of shutting down. It also takes the first exception and propagates it to the GUI exception reporting mechanism (assuming that the GUI is in use, which is not a given).