Add one more situation where we write the cached chain data.
bitcoinx will normally rebuild the chain state from the Genesis block every time. We
have a system for writing the chain data to disk on application shutdown. If someone
always kills the application and it does not shut down properly, this can result in
a long application lock up on start up while bitcoinx rebuilds the chain data from the
Genesis block.

This commit writes the chain data when initial synchronisation of headers to the tip on
any header server is complete, which should ensure that we capture both any processing
of outstanding unprocessed headers and any new synchronised headers on application
startup.