Update MacOS for building the installer using Python 3.9.16.
The reason we are doing this is that the signed builds are done on MacOS 13.3
and this introduces linkage against things that are not available for earlier
versions of MacOS. See the following Python development link for details:

https://github.com/python/cpython/issues/97897

The earliest release of Python with the fix is 3.9.15, with the latest 3.9
release at this time being 3.9.16. So we have to move to that, to solve have
this problem solved for us.