The goal of this work was to set us on the starting point of transitioning to a payment-based experience, not a transaction-based experience. What this means is that the history list is now showing payments not transactions.
There is still a lot of work to do. The starting point was:
- Support creating a new wallet and restoring the stock regtest wallet.
- Have a working history list (albeit with broken ordering).
- Have passing linting, typing and unit tests (albeit with known xfails).

Again, there is a lot of work to do pulling the bits and pieces together.
We need to find all the `TODO(1.4.0) Payments` references and `TODO(nocheckin)`
references and address them all.

Exciting times! Then we can continue with the backup and sync support.