The Azure Pipeline CI is now failing when it tries to build and install hidapi version 0.10.0 (from source because there is no wheel for python 3.9) (#1116)
I have reproduced this locally on both virtualbox and wsl linux when I try to install the deterministic requirements.
After installing cython==0.29.36 and the wheel package, hidapi is able to once again be built and installed.

In terms of the underlying cause - there were no issues on July 16th and then the next time the CI pipeline ran, this issue arose. This is consistent with
the cause being due to a Cython 3.x incompatibility which has been noted by the hidapi maintainers and fixed 5 days ago: https://github.com/trezor/cython-hidapi/issues/155
However, this fix won't be included in the 0.10.0 version so we need to maintain a pin on our Cython version.